requirements.txt
neo4j>=5.20.0
python-dateutil>=2.9.0.post0
# FAISS embedding index (scripts/build_embeddings_faiss.py)
numpy>=1.26.0
pyarrow>=14.0.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
torch>=2.0.0
tqdm>=4.66.0
lightgbm>=4.0.0
